move shared eval function
